requirements.txt
numpy==1.21.2 
scipy==1.8.0
pandas==1.3.4
geopandas==0.10.2
networkx==2.7.1
iso3166==2.1.1
python-dateutil==2.8.2
pycountry-convert==0.7.2 
topojson==1.3
mapclassify==2.4.3
matplotlib==3.5.
shapely==1.8.0
tldextract==3.3.1
